requirements.txt
flask
virustotal3
requests
pefile
capstone
python-magic
flask_sqlalchemy
flask-bootstrap
